Adicionei captação de erros, porém não sei se esta funcionando como deveria, irei testar.

+++ src/main/java/com/supermarket/supermarket/service/MarketService.java
@@ -2,6 +2,7 @@ package com.supermarket.supermarket.service;
 //logica do nosso negocio
 
 import com.supermarket.supermarket.entity.Market;
+import com.supermarket.supermarket.exceptions.ResourceNotFoundException;
 import com.supermarket.supermarket.repository.MarketRepository;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
@@ -33,13 +34,19 @@ public class MarketService {
 
     }
 
-
-
     public List<Market> update(Market market){
+        if(market.getId() == null || !marketRepository.existsById(market.getId())){
+            throw new ResourceNotFoundException("Produto com o id: " + market.getId() + " não encontrado para atualização.");
+        }
+
         marketRepository.save(market);
         return list();
     }
+
     public List<Market> delete(Long id){
+        if(!marketRepository.existsById(id)){
+            throw new ResourceNotFoundException("Produto com o id: " + id + " não encontrado.");
+        }
         marketRepository.deleteById(id);
         return list();
     }
